+++ wol_forward_app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v1.2.0 
+
+Improvements:
+- Security: added AppArmor as an extra protection layer
+- Added parmeters ha_api_url and loglevel to config endpoint data
+
+Changes HTTP Server:
+- Replaced http_api_enabled option by http_api_expose option. 
+- The HTTP server runs always now, but listens to the internal HA network only by default.
+- In normal operation the HTTP API logging (Flask/Werkzeug is chatty on INFO) is adjusted to WARNING when the selected logger level is INFO 
+
+Maintenance:
+- Updated container base image to python:3.12-slim
+
 ## v1.1.2 
 
 Improvements: